added dialog that renders before removing tasks

+++ components/board/display-task.tsx
@@ -2,9 +2,11 @@ import { deleteTask } from "@/app/actions/delete-task";
 import { useState } from "react";
 import { MdClose } from "react-icons/md";
 import { toast } from "sonner";
+import { Alert } from "../alert-dialog/alert";
 
 export const DisplayTask = ({ task }: any) => {
-  const [delTask, setDelTask] = useState<{
+  const [isAlertOpen, setIsAlertOpen] = useState(false);
+  const [taskToDelete, setTaskToDelete] = useState<{
     taskId: number | null;
     boardId: number | null;
     columnId: number | null;
@@ -14,18 +16,31 @@ export const DisplayTask = ({ task }: any) => {
     columnId: null,
   });
 
-  const handleDeleteTask = async (data: {
+  const handleDeleteClick = (data: {
     taskId: number;
     boardId: number;
     columnId: number;
   }) => {
+    setTaskToDelete(data);
+    setIsAlertOpen(true);
+  };
+
+  const handleConfirmDelete = async () => {
+    if (
+      !taskToDelete.taskId ||
+      !taskToDelete.boardId ||
+      !taskToDelete.columnId
+    ) {
+      return;
+    }
+
     try {
-      setDelTask(data);
       const delResponse = await deleteTask(
-        data.taskId,
-        data.boardId,
-        data.columnId
+        taskToDelete.taskId,
+        taskToDelete.boardId,
+        taskToDelete.columnId
       );
+
       if ("success" in delResponse && delResponse.success) {
         toast("Task deleted successfully!", {
           style: {
@@ -43,22 +58,32 @@ export const DisplayTask = ({ task }: any) => {
         "Failed to delete task: " +
           (err instanceof Error ? err.message : "Unknown error")
       );
+    } finally {
+      setIsAlertOpen(false);
     }
   };
 
   return (
-    <div className="flex justify-between items-center shadow-md p-4 rounded-md bg-white">
-      <div className="text-base">{task.title}</div>
-      <MdClose
-        className="cursor-pointer"
-        onClick={() =>
-          handleDeleteTask({
-            taskId: task.id,
-            boardId: task.boardId,
-            columnId: task.columnId,
-          })
-        }
+    <>
+      <div className="flex justify-between items-center shadow-md p-4 rounded-md bg-white">
+        <div className="text-base">{task.title}</div>
+        <MdClose
+          className="cursor-pointer"
+          onClick={() =>
+            handleDeleteClick({
+              taskId: task.id,
+              boardId: task.boardId,
+              columnId: task.columnId,
+            })
+          }
+        />
+      </div>
+
+      <Alert
+        isOpen={isAlertOpen}
+        onClose={() => setIsAlertOpen(false)}
+        onConfirm={handleConfirmDelete}
       />
-    </div>
+    </>
   );
 };
